Support MSFT VS2017 build in ArmPlatformPkg
1. Add MSFT ARM64 assembly files.
2. Update INF files to support MSFT build.
3. Fix C compile warning in VS2017.

+++ ArmPlatformPkg/Drivers/PL011Uart/PL011Uart.c
@@ -190,7 +190,7 @@ PL011UartInitializePort (
       return RETURN_INVALID_PARAMETER;
     }
 
-    Divisor = (UartClkInHz * 4) / *BaudRate;
+    Divisor = (UartClkInHz * 4) / (UINT32) *BaudRate;
     Integer = Divisor >> FRACTION_PART_SIZE_IN_BITS;
     Fractional = Divisor & FRACTION_PART_MASK;
   }
